moving isAdmin to hook
