+++ lib/hooks/index.ts
@@ -1,3 +1,4 @@
+export { default as useIsAdmin } from "./useIsAdmin";
 export { default as useStorage } from "./useStorage";
 export { default as useUserSession } from "./useUserSession";
 export { default as useWindowWidth } from "./useWindowWidth";
